Fix var naming from verb to noun
This commit changes the variable name in the Connections struct from
connectToStoryNodeNamed to connectedStoryNodeName as to avoid confusion
of variables and method names. As a new Swift learner, I had to do a
double take of the code to realize that a method wasn't being called in
the StoryNode class. It is common practice to use verbs for method names
so a verb named variable can cause confusion.